# 2.X版本_OpenAPI.json
{
  "swagger": "2.0",
  "info": {
    "description": "# swagger-bootstrap-ui-demo RESTful APIs",
    "version": "1.0",
    "termsOfService": "http://www.xx.com/",
    "contact": {
      "name": "xx@qq.com"
    }
  },
  "host": "api-frp.ipv4.name",
  "basePath": "/dfm",
...
  "paths": {
...
    "/dfm/data-model/export/{databaseId}": {
      "post": {
        "tags": [
          "DataModel接口"
        ],
        "summary": "导出数据",
        "operationId": "exportUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "databaseId",
            "in": "path",
            "description": "databaseId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/dfm/data-model/import/{databaseId}/{tableName}": {
      "post": {
        "tags": [
          "DataModel接口"
        ],
        "summary": "导入数据",
        "operationId": "importDataUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "databaseId",
            "in": "path",
            "description": "databaseId",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "req",
            "description": "req",
            "required": true,
            "schema": {
              "$ref": "#/definitions/IndexBO",
              "originalRef": "IndexBO"
            }
          },
          {
            "name": "tableName",
            "in": "path",
            "description": "tableName",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
...
    "/dfm/data-model/tables/{databaseId}/{tableName}/indexes": {
      "get": {
        "tags": [
          "DataModel接口"
        ],
        "summary": "获取表索引",
        "operationId": "indexesListUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "databaseId",
            "in": "path",
            "description": "databaseId",
            "required": true,
            "type": "string"
          },
          {
            "name": "tableName",
            "in": "path",
            "description": "tableName",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/IndexBO",
                "originalRef": "IndexBO"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/IndexBO",
                "originalRef": "IndexBO"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": [
          "DataModel接口"
        ],
        "summary": "创建索引",
        "operationId": "rolesCreateUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "databaseId",
            "in": "path",
            "description": "databaseId",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "req",
            "description": "req",
            "required": true,
            "schema": {
              "$ref": "#/definitions/IndexBO",
              "originalRef": "IndexBO"
            }
          },
          {
            "name": "tableName",
            "in": "path",
            "description": "tableName",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/dfm/data-model/tables/{databaseId}/{tableName}/indexes/{indexName}": {
      "delete": {
        "tags": [
          "DataModel接口"
        ],
        "summary": "删除索引",
        "operationId": "indexesDeleteUsingDELETE",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "databaseId",
            "in": "path",
            "description": "databaseId",
            "required": true,
            "type": "string"
          },
          {
            "name": "indexName",
            "in": "path",
            "description": "indexName",
            "required": true,
            "type": "string"
          },
          {
            "name": "tableName",
            "in": "path",
            "description": "tableName",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false
      }
    },
    "/dfm/data-model/tables/{databaseId}/{tableName}/structure": {
      "get": {
        "tags": [
          "DataModel接口"
        ],
        "summary": "获取表结构",
        "operationId": "tablesStructureUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "databaseId",
            "in": "path",
            "description": "databaseId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "tableName",
            "in": "path",
            "description": "tableName",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/TableBO",
                "originalRef": "TableBO"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/TableBO",
                "originalRef": "TableBO"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
...
    "/dfm/data-model/tasks": {
      "get": {
        "tags": [
          "DataModel接口"
        ],
        "summary": "获取导入/导出任务状态",
        "operationId": "tasksUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/TaskBO",
                "originalRef": "TaskBO"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/TaskBO",
                "originalRef": "TaskBO"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
...
  },
  "definitions": {
...
    "IndexBO": {
      "type": "object",
      "properties": {
        "createDate": {
          "type": "string",
          "format": "date-time"
        },
        "createdBy": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "lastLogin": {
          "type": "string",
          "format": "date-time"
        },
        "password": {
          "type": "string"
        },
        "role": {
          "type": "string"
        },
        "roleId": {
          "type": "integer",
          "format": "int64"
        },
        "status": {
          "type": "string"
        },
        "updateDate": {
          "type": "string",
          "format": "date-time"
        },
        "updatedBy": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      },
      "title": "IndexBO"
    },
...
    "TableBO": {
      "type": "object",
      "properties": {
        "createDate": {
          "type": "string",
          "format": "date-time"
        },
        "createdBy": {
          "type": "string"
        },
        "database": {
          "type": "string"
        },
        "databaseId": {
          "type": "integer",
          "format": "int32"
        },
        "fields": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "indexes": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "rows": {
          "type": "string"
        },
        "size": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "updateDate": {
          "type": "string",
          "format": "date-time"
        },
        "updatedBy": {
          "type": "string"
        }
      },
      "title": "TableBO"
    },
    "TaskBO": {
      "type": "object",
      "properties": {
        "createDate": {
          "type": "string",
          "format": "date-time"
        },
        "createdBy": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "lastLogin": {
          "type": "string",
          "format": "date-time"
        },
        "password": {
          "type": "string"
        },
        "role": {
          "type": "string"
        },
        "roleId": {
          "type": "integer",
          "format": "int64"
        },
        "status": {
          "type": "string"
        },
        "updateDate": {
          "type": "string",
          "format": "date-time"
        },
        "updatedBy": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      },
      "title": "TaskBO"
    },
...
  }
}